refactor(CRaterRequest): Create AbstractCRaterRequest (#289)
